삭제 확인창, 삭제 후 화면 갱신

+++ src/pages/memo/MemoItem.jsx
@@ -40,13 +40,16 @@ const MemoItem = ({ item }) => {
     const closePopup = () => {
         setIsOpen(false);
     };
-
     const editMemo = (e) => {
-        openPopup();
-    }
+        openPopup(); // 수정 팝업창 띄우기
+    };
     const deleteMemo = (e) => {
-        deleteData(item.id);
-    }
+        const doDelete = window.confirm('정말로 삭제하시겠습니까?'); // 삭제 확인 메시지 띄우기
+        if(doDelete){
+            deleteData(item.id);
+            window.location.reload();
+        }
+    };
 
   return (
     <Card >
